requirements.txt
streamlit==1.18.1
altair==4.2.2
click==8.0.0
google-cloud-language==2.13.3
python-dotenv
youtube-transcript-api
protobuf<4
google-generativeai
